requirements.txt
asgiref==3.11.0
bcrypt==5.0.0
Django==6.0
psycopg==3.3.2
psycopg-binary==3.3.2
sqlparse==0.5.4
typing_extensions==4.15.0
Pillow==11.0.0
